// src/shared/types/movie.types.ts
import { TMaterialIconName } from './icon.types';

export interface IGenre {
	_id: string;
	name: string;
	slug: string;
	description: string;
	icon: TMaterialIconName;
}

export interface IParameters {
	year: number;
	duration: number;
	country: string;
}

export interface IActor {
	_id: string;
	photo: string;
	name: string;
	countMovies: number;
	slug: string;
}

export interface IMovie {
	_id: string;
	poster: string;
	bigPoster: string;
	title: string;
	parameters: IParameters;
	genres: IGenre[];
	actors: IActor[];
	countOpened: number;
	videoUrl: string;
	rating: number;
	slug: string;
}